refactor: split dropdown into 2 components
